// src/app/shared/models/data-vizualization.ts
// service-health.model.ts
export interface ServiceVersion {
    app?: string;
    commit: string;
    branch: string;
    tag: string;
    buildTime?: string;
  }
  
  export interface MemoryUsage {
    rss: string;
    heapTotal: string;
    heapUsed: string;
    external: string;
  }
  
  export interface MemoryHealthDetails {
    heapUsedPercent: number;
    threshold: number;
    heapUsed: number;
    heapTotal: number;
  }
  
  export interface MemoryHealthCheck {
    status: 'HEALTHY' | 'DEGRADED' | 'UNHEALTHY';
    details: MemoryHealthDetails;
    responseTime: string;
  }
  
  export interface DataVisualization {
    status: 'HEALTHY' | 'DEGRADED' | 'UNHEALTHY' | 'UNKNOWN';
    timestamp: string;
    service: string;
    version: Record<string, ServiceVersion>;
    runtime: {
      nodeVersion: string;
      platform: string;
      arch: string;
      uptime: string;
      uptimeMs: number;
      memoryUsage: MemoryUsage;
      environment: string;
      domainEnv: string;
    };
    checks: {
      memoryHealthCheck: MemoryHealthCheck;
    };
  }
  